requirements.txt
openai>=1.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
streamlit>=1.28.0
plotly>=5.0.0